Se crea un nuevo controller y service para unificar la funcionalidad del control del stock.
