+++ client/html/administrador.html
@@ -131,7 +131,7 @@
               </div>
             </div>
             <div class="col-md-12">
-              <a href="" class="btn btn-dark btn-agregar-articulo">AGREGAR</a>
+              <a href="#" class="btn btn-dark btn-agregar-articulo">AGREGAR</a>
             </div>
           </div>
         </div>
