+++ client/js/articulo_detalle_producto.js
@@ -14,7 +14,7 @@ let imagenPrincipal = document.querySelector("#img-principal");
 async function load() {
     try {
         let params = processParams();
-        const URL = `/articulo/${params["categoria"]}/${params["index"]}`;
+        const URL = `/stock/${params["categoria"]}/${params["index"]}`;
         let response = await fetch(URL);
         if (response.ok) {
             let articulo = await response.json();
@@ -53,7 +53,7 @@ function cargarImagenes(i, imagen) {
  **/
 // Traigo los artículos según la categoría.
 async function cargarArticulos(categoria) {
-    const URL = `/articulo/${categoria}`;
+    const URL = `/stock/${categoria}`;
     try {
         let response = await fetch(URL);
         if (response.ok) {
