+++ src/app.module.ts
@@ -5,13 +5,15 @@ import { ServeStaticModule } from '@nestjs/serve-static';
 import { join } from 'path';
 import { NegocioController } from './negocio/negocio.controller';
 import { NegocioService } from './negocio/negocio.service';
+import { StockController } from './stock/stock.controller';
+import { StockService } from './stock/stock.service';
 
 @Module({
   imports: [ServeStaticModule.forRoot({
     rootPath: join(__dirname, '..',
     'client'),
     }),],
-  controllers: [AppController, NegocioController],
-  providers: [AppService, NegocioService],
+  controllers: [AppController, NegocioController, StockController],
+  providers: [AppService, NegocioService, StockService],
 })
 export class AppModule {}
