+++ src/stock/stock.controller.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { StockController } from './stock.controller';
+
+describe('StockController', () => {
+  let controller: StockController;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      controllers: [StockController],
+    }).compile();
+
+    controller = module.get<StockController>(StockController);
+  });
+
+  it('should be defined', () => {
+    expect(controller).toBeDefined();
+  });
+});
