+++ src/stock/stock.service.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { StockService } from './stock.service';
+
+describe('StockService', () => {
+  let service: StockService;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [StockService],
+    }).compile();
+
+    service = module.get<StockService>(StockService);
+  });
+
+  it('should be defined', () => {
+    expect(service).toBeDefined();
+  });
+});
